requirements.txt
discord-py-interactions>=5.13.1
FactorioAPI>=1.1.2
orjson>=3.10.6
requests==2.32.3